fix(aes): Fix undefined behavior in AES key expansion and improve bcrypt KDF
- Cast S-box lookups to uint32_t before shifting to prevent signed integer overflow UB
- Added bcrypt_pbkdf implementation for proper key derivation from passwords
- Fixed salt length handling in deriveKeyFromParams (use actual converted length)
- Added missing IV override when -v flag is provided
- Replaced complex base64+cipher chaining with clean pipe/fork implementation
- Added secureZeroMemory for password buffers in prompt
- Fixed bcrypt function signatures and add missing includes
- Improved error handling in key derivation functions

+++ src/cipher/aes/transform.c
@@ -37,16 +37,16 @@ uint32_t aesExpandKey(const uint8_t *key, size_t keyLen, uint32_t roundKeys[60])
 	for (i = keyLen / 4; i < 4 * (nbRounds + 1); i++) {
 		temp = roundKeys[i-1];
 		if (i % (keyLen / 4) == 0) {
-			temp = (g_aes_sbox[(temp >> 16) & 0xFF] << 24) |
-				   (g_aes_sbox[(temp >>  8) & 0xFF] << 16) |
-				   (g_aes_sbox[(temp	  ) & 0xFF] <<  8) |
-				   (g_aes_sbox[(temp >> 24) & 0xFF]);
+			temp = ((uint32_t)g_aes_sbox[(temp >> 16) & 0xFF] << 24) |
+				   ((uint32_t)g_aes_sbox[(temp >>  8) & 0xFF] << 16) |
+				   ((uint32_t)g_aes_sbox[(temp      ) & 0xFF] <<  8) |
+				   ((uint32_t)g_aes_sbox[(temp >> 24) & 0xFF]);
 			temp ^= rcon[i / (keyLen / 4)];
 		} else if ((keyLen / 4) == 8 && i % 8 == 4) {
-			temp = (g_aes_sbox[(temp >> 24) & 0xFF] << 24) |
-				   (g_aes_sbox[(temp >> 16) & 0xFF] << 16) |
-				   (g_aes_sbox[(temp >>  8) & 0xFF] <<  8) |
-				   (g_aes_sbox[(temp	  ) & 0xFF]);
+			temp = ((uint32_t)g_aes_sbox[(temp >> 24) & 0xFF] << 24) |
+				   ((uint32_t)g_aes_sbox[(temp >> 16) & 0xFF] << 16) |
+				   ((uint32_t)g_aes_sbox[(temp >>  8) & 0xFF] <<  8) |
+				   ((uint32_t)g_aes_sbox[(temp      ) & 0xFF]);
 		}
 		roundKeys[i] = roundKeys[i - (keyLen / 4)] ^ temp;
 	}
